Adding verification on reviews and users

+++ part3/app/api/v1/reviews.py
@@ -1,4 +1,5 @@
 from flask_restx import Namespace, Resource, fields
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from app.services import facade
 
 api = Namespace('reviews', description='Review operations')
@@ -23,6 +24,7 @@ class ReviewList(Resource):
     @api.expect(review_model)
     @api.response(201, 'Review successfully created')
     @api.response(400, 'Invalid input data')
+    @jwt_required()
     def post(self):
         """
         Register a new review.
@@ -48,6 +50,19 @@ class ReviewList(Resource):
         existing_place = facade.get_place(review_data['place_id'])
         if not existing_place:
             return {'error': 'This place doesn\'t exist'}, 404
+  
+        current_user = get_jwt_identity()
+        if current_user != review_data['user_id']:
+            return {'error': 'Unauthorized action'}, 403
+
+        if current_user == existing_place.owner_id:
+            return {"message": "You cannot review your own place."}, 400
+
+        for existing_review in existing_place.reviews:
+            if current_user == existing_review.user_id:
+                return {
+                    "message": "You have already reviewed this place."
+                    }, 400
 
         try:
             review = facade.create_review(review_data)
@@ -122,6 +137,7 @@ class ReviewResource(Resource):
     @api.response(200, 'Review updated successfully')
     @api.response(404, 'Review not found')
     @api.response(400, 'Invalid input data')
+    @jwt_required()
     def put(self, review_id):
         """
         Update an existing review.
@@ -145,13 +161,14 @@ class ReviewResource(Resource):
             return {'error': 'Review not found'}, 404
         update_data = api.payload
 
-        existing_user = facade.get_user(update_data['user_id'])
-        if not existing_user:
-            return {'error': 'This user doesn\'t exist'}, 404
+        for key in update_data:
+            if key == 'place_id' or key == 'user_id' or key == 'id':
+                if update_data[key] != getattr(review, key):
+                    return {'error': 'You cannot modify ids.'}, 403
 
-        existing_place = facade.get_place(update_data['place_id'])
-        if not existing_place:
-            return {'error': 'This place doesn\'t exist'}, 404
+        current_user = get_jwt_identity()
+        if current_user != review.user_id:
+            return {"error": "Unauthorized action."}, 403
 
         try:
             facade.update_review(review_id, update_data)
@@ -162,6 +179,7 @@ class ReviewResource(Resource):
 
     @api.response(200, 'Review deleted successfully')
     @api.response(404, 'Review not found')
+    @jwt_required()
     def delete(self, review_id):
         """
         Delete a review.
@@ -181,6 +199,10 @@ class ReviewResource(Resource):
         review = facade.get_review(review_id)
         if not review:
             return {'error': 'Review not found'}, 404
+
+        current_user = get_jwt_identity()
+        if current_user != review.user_id:
+            return {'message': 'This review is not yours'}, 400
 
         review = facade.delete_review(review_id)
         return {"message": "Review deleted successfully"}
